refactor: replace IoDispatcher and DefaultDispatcher with Dispatcher annotations for consistency
