+++ core/data/src/main/java/com/keelim/core/data/source/DefaultTaskRepositoryImpl.kt
@@ -1,7 +1,8 @@
 package com.keelim.core.data.source
 
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import com.keelim.common.di.ApplicationScope
-import com.keelim.common.di.DefaultDispatcher
 import com.keelim.core.data.model.toLocal
 import com.keelim.core.data.model.toNetwork
 import com.keelim.core.database.dao.TaskDao
@@ -20,7 +21,7 @@ import javax.inject.Inject
 class DefaultTaskRepositoryImpl @Inject constructor(
     private val localDataSource: TaskDao,
     private val networkDataSource: TaskNetworkDataSource,
-    @DefaultDispatcher private val dispatcher: CoroutineDispatcher,
+    @Dispatcher(KeelimDispatchers.IO) private val dispatcher: CoroutineDispatcher,
     @ApplicationScope private val scope: CoroutineScope,
 ) : DefaultTaskRepository {
