+++ core/data/src/main/java/com/keelim/core/data/source/alarm/AlarmRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package com.keelim.core.data.source.alarm
 
-import com.keelim.common.di.IoDispatcher
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import com.keelim.core.database.dao.AlarmDao
 import com.keelim.core.database.mapper.toAlarm
 import com.keelim.core.database.mapper.toAlarmEntity
@@ -15,7 +16,7 @@ import javax.inject.Inject
 
 class AlarmRepositoryImpl @Inject constructor(
     private val localDataSource: AlarmDao,
-    @IoDispatcher private val io: CoroutineDispatcher,
+    @Dispatcher(KeelimDispatchers.IO) private val io: CoroutineDispatcher,
 ) : AlarmRepository {
     override suspend fun insertAlarm(
         title: String,
