+++ core/data/src/main/java/com/keelim/core/data/source/prompt/PromptRepositoryImpl.kt
@@ -1,7 +1,8 @@
 package com.keelim.core.data.source.prompt
 
 import com.google.ai.client.generativeai.GenerativeModel
-import com.keelim.common.di.IoDispatcher
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import com.keelim.data.repository.PromptRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
@@ -10,7 +11,7 @@ import javax.inject.Inject
 
 class PromptRepositoryImpl @Inject constructor(
     val generativeModel: GenerativeModel,
-    @IoDispatcher val dispatcher: CoroutineDispatcher,
+    @Dispatcher(KeelimDispatchers.IO) val dispatcher: CoroutineDispatcher,
 ) : PromptRepository {
     override suspend fun getContent(prompt: String): Result<String> {
         return runCatching {
